Move code to better place
